Function Map should be sorted in decreasing order
We want the wrapping function returned by the functionmap consumer to
contain the closest function to the source position. In order to do
this, we will sort the function map by start source position in
decreasing order so that the first function we encounter that wraps the
source position will be the closest function to the source position.